rewrite compressor so we can actually use it
this makes some changes to the compression so that it actually works with lists of keys, instead of one big string, which is what we'll need in the final version. it also adds one optimization for no-ops, and removes the end-of-data marker, since in the parser, we'll have the number of keys handed to us, too -- which makes a separate end-of-data marker unnecessary.